terraform_agent/pyproject.toml
[project]
name = "tf-migrator"
version = "0.1.0"
description = "AI-powered Terraform cloud migration agent"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "langgraph>=0.2.0",
    "llama-index>=0.11.0",
    "python-hcl2>=4.3.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "networkx>=3.0",
    "rich>=13.0",
    "typer>=0.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.5.0",
]

[project.scripts]
tf-migrate = "tf_migrator.main:app"
